photo-potato-backend/pyproject.toml
[tool.poetry]
name = "photo-potato-backend"
version = "0.1.0"
description = "Flask backend"
authors = ["chilir <minghao_li@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-SQLAlchemy = "^2.5.1"
psycopg2 = "^2.9.3"
python-dotenv = "^0.21.0"
Flask-Cors = "^3.0.10"
Pillow = "^9.2.0"
Flask-Bcrypt = "^1.0.1"
Flask-JWT-Extended = "^4.4.4"
opencv-python = "^4.6.0"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
mypy = "^0.971"
black = {version = "^22.8.0", allow-prereleases = true}
ipykernel = "^6.15.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
